refactor(ui): use shared FormattedThousandsNumberInput component across different onboarding pages
